feat: move metadata generation to a separate file for better organization
